Move "odo list" to "odo list component"
